+++ app/templatetags/tags.py
@@ -1,23 +1,8 @@
 from django import template
-from wagtail.models import Site
+from app.utils import get_menu
 
 
 register = template.Library()
-
-
-def get_menu(context):
-    """Function for header and footer. Returns menu items"""
-
-    menu = [Site.find_for_request(context["request"]).root_page]
-    [
-        menu.append(item)
-        for item in Site.find_for_request(context["request"])
-        .root_page.get_children()
-        .live()
-        .in_menu()
-    ]
-
-    return menu
 
 
 @register.inclusion_tag("tags/header.html", takes_context=True)
@@ -25,4 +10,12 @@ def get_header(context):
     return {
         "request": context["request"],
         "header_menu": get_menu(context),
+    }
+
+
+@register.inclusion_tag("tags/footer.html", takes_context=True)
+def get_footer(context):
+    return {
+        "request": context["request"],
+        "footer_menu": get_menu(context),
     }
