+++ templates/tags/footer.html
@@ -1,0 +1,26 @@
+{% load wagtailcore_tags tags %}
+
+<div class="footer">
+    <div class="footer__container">
+        
+        <div class="footer__image">
+            <img src="https://cdn.dribbble.com/userupload/3158902/file/original-7c71bfa677e61dea61bc2acd59158d32.jpg?resize=400x0" alt="">
+        </div>
+        <div class="footer__menu">
+            {% for item in footer_menu %}
+                <a href="{% pageurl item %}">{{ item }}</a>
+            {% endfor %}
+        </div>
+        <div class="footer__links">
+            <p>any email</p>
+            <p>any phone number</p>
+            <p>any bot</p>
+        </div>
+        <div class="footer__button">
+            <a href="{% slugurl 'form-page' %}">
+                <button class="contact_us">Contact us</button>
+            </a>
+        </div>
+        
+    </div>
+</div>
